+++ labgrid/remote/exporter.py
@@ -71,26 +71,25 @@ class ResourceExport(ResourceEntry):
         self._stop(self.start_params)
         self.start_params = None
 
-    def need_restart(self):
-        """
-        Check if the previously used start parameters have changed so that a
-        restart is needed.
-        """
-        start_params = self._get_start_params()
-        if self.start_params != start_params:
-            self.logger.info("restart needed (%s -> %s)", self.start_params, start_params)
-            return True
-        return False
-
     def poll(self):
-        dirty = False
         # poll and check for updated params/avail
         self.local.poll()
-        if self.avail != self.local.avail:
-            if self.local.avail:
+
+        if self.local.avail and self.acquired:
+            start_params = self._get_start_params()
+            if self.start_params is None:
                 self.start()
-            else:
+            elif self.start_params != start_params:
+                self.logger.info("restart needed (%s -> %s)", self.start_params, start_params)
                 self.stop()
+                self.start()
+        else:
+            if self.start_params is not None:
+                self.stop()
+
+        # check if resulting information has changed
+        dirty = False
+        if self.avail != self.local.avail:
             self.data['avail'] = self.local.avail
             dirty = True
         params = self._get_params()
@@ -99,12 +98,18 @@ class ResourceExport(ResourceEntry):
         params['extra']['proxy_required'] = self.proxy_required
         params['extra']['proxy'] = self.proxy
         if self.params != params:
-            if self.local.avail and self.need_restart():
-                self.stop()
-                self.start()
             self.data['params'].update(params)  # pylint: disable=unsubscriptable-object
             dirty = True
+
         return dirty
+
+    def acquire(self, *args, **kwargs):
+        super().acquire(*args, **kwargs)
+        self.poll()
+
+    def release(self, *args, **kwargs):
+        super().release(*args, **kwargs)
+        self.poll()
 
 
 @attr.s(cmp=False)
@@ -161,13 +166,15 @@ class USBSerialPortExport(ResourceExport):
         # stop ser2net
         child = self.child
         self.child = None
+        port = self.port
+        self.port = None
         child.terminate()
         try:
             child.wait(1.0)
         except subprocess.TimeoutExpired:
             child.kill()
             child.wait(1.0)
-        self.logger.info("stopped ser2net for %s on port %d", start_params['path'], self.port)
+        self.logger.info("stopped ser2net for %s on port %d", start_params['path'], port)
 
 
 exports["USBSerialPort"] = USBSerialPortExport
@@ -381,6 +388,12 @@ class ExporterSession(ApplicationSession):
         - bail out if we are unsuccessful
         """
         print(details)
+
+        prefix = 'org.labgrid.exporter.{}'.format(self.name)
+        await self.register(self.acquire, '{}.acquire'.format(prefix))
+        await self.register(self.release, '{}.release'.format(prefix))
+        await self.register(self.version, '{}.version'.format(prefix))
+
         try:
             resource_config = ResourceConfig(self.config.extra['resources'])
             for group_name, group in resource_config.data.items():
@@ -391,6 +404,7 @@ class ExporterSession(ApplicationSession):
                         continue
                     cls = params.pop('cls', resource_name)
 
+                    # this may call back to acquire the resource immediately
                     await self.add_resource(
                         group_name, resource_name, cls, params
                     )
@@ -401,11 +415,6 @@ class ExporterSession(ApplicationSession):
             return
 
         self.poll_task = self.loop.create_task(self.poll())
-
-        prefix = 'org.labgrid.exporter.{}'.format(self.name)
-        await self.register(self.acquire, '{}.acquire'.format(prefix))
-        await self.register(self.release, '{}.release'.format(prefix))
-        await self.register(self.version, '{}.version'.format(prefix))
 
     async def onLeave(self, details):
         """Cleanup after leaving the coordinator connection"""
@@ -424,16 +433,14 @@ class ExporterSession(ApplicationSession):
             await asyncio.sleep(0.5) # give others a chance to clean up
         self.loop.stop()
 
-    async def acquire(self, group_name, resource_name):
-        # TODO: perform local actions when a resource is acquired
-        #resource = self.groups[group_name][resource_name]
-        #resource.acquire()
+    async def acquire(self, group_name, resource_name, place_name):
+        resource = self.groups[group_name][resource_name]
+        resource.acquire(place_name)
         await self.update_resource(group_name, resource_name)
 
     async def release(self, group_name, resource_name):
-        # TODO: perform local actions when a resource is released
-        #resource = self.groups[group_name][resource_name]
-        #resource.release()
+        resource = self.groups[group_name][resource_name]
+        resource.release()
         await self.update_resource(group_name, resource_name)
 
     async def version(self):
@@ -451,18 +458,12 @@ class ExporterSession(ApplicationSession):
                     traceback.print_exc()
                     continue
                 if changed:
-                    # resource has changed
-                    data = resource.asdict()
-                    print(data)
-                    await self.call(
-                        'org.labgrid.coordinator.set_resource', group_name,
-                        resource_name, data
-                    )
+                    await self.update_resource(group_name, resource_name)
 
     async def poll(self):
         while True:
             try:
-                await asyncio.sleep(1.0)
+                await asyncio.sleep(0.25)
                 await self._poll_step()
             except asyncio.CancelledError:
                 break